+++ src/app/_mock/assets.ts
@@ -187,46 +187,3 @@ export const owned: ownedAssets[] = [
         fiat_value: '0.000',
     },
 ]
-
-export const transactions: tx[] = [
-    {
-        date: '21-06-2023 07:24',
-        id: '#88hfd02333633',
-        asset: 'Tron - TRX',
-        status: 'PENDING',
-        amt:'0.0002',
-        tx_type: 'BUY',
-    },
-    {
-        date: '21-06-2023 07:24',
-        id: '#88hfd02333633',
-        asset: 'Tron - TRX',
-        status: 'PENDING',
-        amt:'0.0002',
-        tx_type: 'BUY',
-    },
-    {
-        date: '21-06-2023 07:24',
-        id: '#88hfd02333633',
-        asset: 'ETH - USDT',
-        status: 'SUCCESSFUL',
-        amt:'0.0002',
-        tx_type: 'SELL',
-    },
-    {
-        date: '21-06-2023 07:24',
-        id: '#88hfd02333633',
-        asset: 'ETH - USDT',
-        status: 'SUCCESSFUL',
-        amt:'0.0002',
-        tx_type: 'SELL',
-    },
-    {
-        date: '21-06-2023 07:24',
-        id: '#88hfd02333633',
-        asset: 'ETH - USDT',
-        status: 'FAILED',
-        amt:'0.0002',
-        tx_type: 'SELL',
-    },
-]
