+++ src/app/_mock/trancs.ts
@@ -1,0 +1,44 @@
+import {tx} from "@/types";
+
+export const transactions: tx[] = [
+    {
+        date: '21-06-2023 07:24',
+        id: '#88hfd02333633',
+        asset: 'Tron - TRX',
+        status: 'PENDING',
+        amt:'0.0002',
+        tx_type: 'BUY',
+    },
+    {
+        date: '21-06-2023 07:24',
+        id: '#88hfd02333633',
+        asset: 'Tron - TRX',
+        status: 'PENDING',
+        amt:'0.0002',
+        tx_type: 'BUY',
+    },
+    {
+        date: '21-06-2023 07:24',
+        id: '#88hfd02333633',
+        asset: 'ETH - USDT',
+        status: 'SUCCESSFUL',
+        amt:'0.0002',
+        tx_type: 'SELL',
+    },
+    {
+        date: '21-06-2023 07:24',
+        id: '#88hfd02333633',
+        asset: 'ETH - USDT',
+        status: 'SUCCESSFUL',
+        amt:'0.0002',
+        tx_type: 'SELL',
+    },
+    {
+        date: '21-06-2023 07:24',
+        id: '#88hfd02333633',
+        asset: 'ETH - USDT',
+        status: 'FAILED',
+        amt:'0.0002',
+        tx_type: 'SELL',
+    },
+]
