+++ src/app/components/modal/TxDetails.vue
@@ -1,5 +1,5 @@
 <template>
-  <Dialog v-model:visible="visible" modal header="Transaction Details"  class="2xl:w-[45vw] xl:w-[50vw] lg:w-[60vw] md:w-[80vw] w-[90vw]">
+  <Dialog v-model:visible="modal" modal header="Transaction Details"  class="2xl:w-[45vw] xl:w-[50vw] lg:w-[60vw] md:w-[80vw] w-[90vw]">
     <div>
       <p class="pb-2">Transaction Info</p>
       <div class="border rounded flex ">
@@ -23,9 +23,13 @@
 </template>
 
 <script setup lang="ts">
+import { storeToRefs } from "pinia";
+import {useTxStore} from "@/app/stores/main/tx.store";
 
 import Dialog from 'primevue/dialog';
 
+const store = useTxStore();
+const { modal } = storeToRefs(store);
 </script>
 
 <style scoped>
