+++ src/app/stores/main/tx.store.ts
@@ -1,0 +1,11 @@
+import {defineStore} from "pinia";
+
+export const useTxStore = defineStore({
+    id: 'tx',
+    state: () => ({
+        tx_details: Boolean,
+        modal: false
+    }),
+    actions: {},
+    getters: {}
+})
